Nasty hack to allow :for on participant objects
This should be replaced with a better solution. Right now this
allows you to say :for => [:participant_friends] and it will call
self.participant.send(:friends) in the PfeedItem.deliver method.
I needed this to be able to notify all users who have commented
on a particular item, that another item has been created. I did
not want to notify all users, only the voices in that topic.
I am definitely open to alternative methods for accomplishing
this, in the mean time, this works. I initially tried to overwrite
the deliver method in each STI sub PfeedItems model to keep this
code out of the main codebase, however, it would only work once.
For some reason, the second time it was called, it would skip
calling the STI model and revert back to calling the base model
which would break because we're trying to call a method on the
participant.