Fixed bug in PfeedItem.new_pfeed_item conditional
Switched false to nil. Initial conditional falls through after
the first item has been created, because it checks if the custom
class is nil?, but the custom classe is FALSE, so FALSE.nil?
fails the conditional, and we drop through into the default
PfeedItem constructor rather than using the STI klass model as a
constructor.